backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
vaderSentiment==3.3.2
tweepy==4.14.0
praw==7.7.1
google-api-python-client==2.108.0
requests==2.31.0
gunicorn==21.2.0




